+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+2.0.0 version (2019-10-17)
+--------------------------------
+* Breaking API: Some functions are now async. In particular, the top-level
+  `sign_file` function is now an async function.
+* Restructured module layout
+* Added docs
+
 1.3.0 (2019-09-12)
 ------------------